Unet_Depth_Esimation/requirements.txt
# PyTorch and related packages
torch>=1.11.0
torchvision>=0.12.0
torchaudio>=0.11.0

# Core scientific computing
numpy>=1.21.0
matplotlib>=3.5.0

# Image processing
Pillow>=8.3.0
imageio>=2.19.0

# Progress bars and utilities
tqdm>=4.62.0

# Dataset handling
h5py>=3.7.0

# Experiment tracking
wandb>=0.12.0

# Data handling
requests>=2.26.0

# Standard library dependencies (usually included with Python)
# os, sys, shutil, time, random, pathlib, argparse, math are standard library modules
